REC-186: Added numDaysRemaining to job model. Need to update this to be duration of job when status endpoint for jobs is called. Endpoint was added in another branch.

+++ src/main/java/com/avalanche/tmcs/DataLoader.java
@@ -205,7 +205,7 @@ public class DataLoader implements ApplicationRunner {
                 long minGPA = (long) job.get("minGPA");
                 //has work experience
                 //match threshold
-                long duration = (long) job.get("duration");
+                int duration = (int) job.get("duration");
                 String problemStatement = (String) job.get("problemStatement");
                 String video = (String) job.get("video");
